pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promo-query-py"
version = "0.1.0"
description = "Motor consultable de promociones bancarias y precios de combustibles para Paraguay."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12",
  "fastapi>=0.115",
  "Jinja2>=3.1",
  "lxml>=5.2",
  "pdfplumber>=0.11",
  "psycopg[binary]>=3.2",
  "pypdf>=5.4",
  "python-multipart>=0.0.9",
  "pydantic>=1.10,<3",
  "PyYAML>=6.0",
  "requests>=2.32",
  "SQLAlchemy>=2.0",
  "typer>=0.12",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28",
  "pytest>=8.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
